+++ v3/crates/execute/src/ir/order_by.rs
@@ -113,7 +113,7 @@ pub(crate) fn build_ndc_order_by_element<'s>(
         // a relationship column, we'll have to join all the paths to specify NDC,
         // what relationships needs to be traversed to access this column
         Annotation::Input(InputAnnotation::Model(
-            schema::ModelInputAnnotation::ModelOrderByArgument { ndc_column },
+            schema::ModelInputAnnotation::ModelOrderByArgument { ndc_column, .. },
         )) => {
             let order_by_value = argument.value.as_enum()?;
             let order_direction = match &order_by_value.info.generic {